fix: fixed lazy loading of modules
